refactor(forja): header ÚNICO (ForjaHub) em todas as abas + remove eyebrow
Wagner: "page header em todos iguais? remova o Desenvolvimento · MCP · Projeção do git".

- ForjaHub: novo componente = header único (título Forja + ações sino/⌘K/Novo issue
  + tab-strip de 9 abas). SEM eyebrow.
- Cockpit: usa <ForjaHub> (removido eyebrow + header/tab-strip inline duplicados).
- Telas absorvidas (Team/Tasks/CcSessions/Scorecard): renderizam <ForjaHub active=...>
  no topo → header IDÊNTICO em todas as abas do hub (não mais cada tela com header próprio).

Agora /forja e /team-mcp/* mostram o mesmo header Forja + tab-strip. eslint só com
warning a11y pré-existente (Tasks). Refs: FORJA fusão · header único
