+++ resources/js/Pages/team-mcp/Scorecard/Index.tsx
@@ -15,6 +15,7 @@ import { useEffect, type ReactNode } from 'react';
 import { AlertCircle, CheckCircle2, RefreshCw } from 'lucide-react';
 import { Button } from '@/Components/ui/button';
 import { PageHeader } from '@/Components/PageHeader';
+import ForjaHub from '@/Pages/team-mcp/Forja/_components/ForjaHub';
 import KpiGrid from '@/Components/shared/KpiGrid';
 import KpiCard from '@/Components/shared/KpiCard';
 import { cn } from '@/Lib/utils';
@@ -76,6 +77,7 @@ function ScorecardIndex({ facts, checks, meta }: Props) {
 
   return (
     <>
+      <ForjaHub active="saude" />
       <PageHeader
         title="Saúde do MCP"
         subtitle={`Facts + Checks · janela ${meta.period_days}d · fonte ${meta.source}`}
